requirements.txt
Flask==2.2.0
pandas==2.0.3
openpyxl==3.1.2
pillow==10.3.0
img2pdf==0.5.1
Werkzeug==2.2.0